+++ src/storage/store.rs
@@ -15,6 +15,11 @@ pub use key::{StoreKey, StoreKeyError, StoreKeys};
 pub use memory::MemoryStore;
 pub use prefix::{StorePrefix, StorePrefixError, StorePrefixes};
 
+#[cfg(feature = "http")]
+mod http;
+#[cfg(feature = "http")]
+pub use http::{HTTPStore, HTTPStoreCreateError};
+
 // pub use store_plugin::{StorePlugin, StorePluginCreateError}; // Currently disabled.
 
 use std::sync::Arc;
@@ -49,10 +54,10 @@ pub type ReadableWritableStore = Arc<dyn ReadableWritableStoreExtension>;
 // pub type ReadableWritableStorePlugin = StorePlugin<ReadableWritableStore>;
 // inventory::collect!(ReadableWritableStorePlugin);
 
-/// A store with [`Send`], [`Sync`] and an optional [`uri_scheme`](StoreExtension::uri_scheme).
+/// Traits for a store extension.
 pub trait StoreExtension: Send + Sync {
-    /// The URI scheme of the store, if it has one.
-    fn uri_scheme(&self) -> Option<&'static str>;
+    // /// The URI scheme of the store, if it has one.
+    // fn uri_scheme(&self) -> Option<&'static str>;
 }
 
 /// A readable store extension.